[Bug]-fixed 단어 인식시 동일한 글자가 있을경우 좌표를 제대로 인식하지 못하는 버그 수정
